+++ spec/functions/lookup_spec.rb
@@ -1,141 +1,257 @@
 require 'spec_helper'
+require 'mock_vault_helper'
 
+include PuppetVaultLookupHelpers
 describe 'vault_lookup::lookup' do
   let(:function) { subject }
-
-  let(:auth_success_data) do
-    <<~JSON
-      {
-        "request_id": "03d11bd4-b994-c432-150f-5703a75641d1",
-        "lease_id": "",
-        "renewable": false,
-        "lease_duration": 0,
-        "data": null,
-        "wrap_info": null,
-        "warnings": null,
-        "auth": {
-          "client_token": "7dad29d2-40af-038f-cf9c-0aeb616f8d20",
-          "accessor": "fd0c3269-9642-25e5-cebe-27a732be53a0",
-          "policies": [
-            "default",
-            "secret_reader"
-          ],
-          "token_policies": [
-            "default",
-            "secret_reader"
-          ],
-          "metadata": {
-            "authority_key_id": "b7:da:18:2f:cc:09:18:5d:d0:c5:24:0a:0a:66:46:ba:0d:f0:ea:4a",
-            "cert_name": "vault.docker",
-            "common_name": "localhost",
-            "serial_number": "5",
-            "subject_key_id": "ea:00:c0:0b:2d:38:01:28:ba:16:1f:08:64:de:0a:7c:8f:b7:43:33"
-          },
-          "lease_duration": 604800,
-          "renewable": true,
-          "entity_id": "e1bc06c5-303e-eec7-bf58-2a74fae2ec3d"
-        }
-      }
-    JSON
-  end
-
-  let(:auth_failure_data) do
-    '{"errors":["invalid certificate or no client certificate supplied"]}'
-  end
-
-  let(:secret_success_data) do
-    '{"request_id":"e394e8ef-78f3-ac85-fbeb-33f060e911d4","lease_id":"","renewable":false,"lease_duration":604800,"data":{"foo":"bar"},"wrap_info":null,"warnings":null,"auth":null}
-'
-  end
-
-  let(:permission_denied_data) do
-    '{"errors":["permission denied"]}'
-  end
 
   it 'errors for malformed uri' do
     expect {
       function.execute('/v1/whatever', 'vault.docker')
     }.to raise_error(Puppet::Error, %r{Unable to parse a hostname})
+
+    expect {
+      function.execute('/v1/whatever', 'vault_addr' => 'vault.docker')
+    }.to raise_error(Puppet::Error, %r{Unable to parse a hostname})
   end
 
-  it 'errors when no vault_url set and no VAULT_ADDR environment variable' do
+  it 'errors when no vault_addr set and no VAULT_ADDR environment variable' do
     expect {
       function.execute('/v1/whatever')
-    }.to raise_error(Puppet::Error, %r{No vault_url given and VAULT_ADDR env variable not set})
+    }.to raise_error(Puppet::Error, %r{No vault_addr given and VAULT_ADDR env variable not set})
+
+    expect {
+      function.execute('/v1/whatever', {})
+    }.to raise_error(Puppet::Error, %r{No vault_addr given and VAULT_ADDR env variable not set})
   end
 
   it 'raises a Puppet error when auth fails' do
-    connection = instance_double('Puppet::Network::HTTP::Connection', address: 'vault.doesnotexist')
-    expect(Puppet::Network::HttpPool).to receive(:http_instance).and_return(connection)
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthFailure)
+    vault_server.start_vault do |port|
+      expect {
+        function.execute('thepath', "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 403 response code from vault.*invalid certificate or no client certificate supplied})
 
-    response = Net::HTTPForbidden.new('1.1', 403, auth_failure_data)
-    allow(response).to receive(:body).and_return(auth_failure_data)
-    expect(connection).to receive(:post).with('/v1/auth/cert/login', '').and_return(response)
-
-    expect {
-      function.execute('thepath', 'https://vault.doesnotexist:8200')
-    }.to raise_error(Puppet::Error, %r{Received 403 response code from vault.*invalid certificate or no client certificate supplied})
+      expect {
+        function.execute('thepath', 'vault_addr' => "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 403 response code from vault.*invalid certificate or no client certificate supplied})
+    end
   end
 
   it 'raises a Puppet error when data lookup fails' do
-    connection = instance_double('Puppet::Network::HTTP::Connection', address: 'vault.doesnotexist')
-    expect(Puppet::Network::HttpPool).to receive(:http_instance).and_return(connection)
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupDenied)
+    vault_server.start_vault do |port|
+      expect {
+        function.execute('kv/test', "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 403 response code from vault at .* for secret lookup.*permission denied})
 
-    auth_response = Net::HTTPOK.new('1.1', 200, '')
-    expect(auth_response).to receive(:body).and_return(auth_success_data)
-    expect(connection).to receive(:post).with('/v1/auth/cert/login', '').and_return(auth_response)
-
-    secret_response = Net::HTTPForbidden.new('1.1', 403, permission_denied_data)
-    allow(secret_response).to receive(:body).and_return(permission_denied_data)
-    expect(connection)
-      .to receive(:get)
-      .with('/v1/secret/test', hash_including('X-Vault-Token' => '7dad29d2-40af-038f-cf9c-0aeb616f8d20'))
-      .and_return(secret_response)
-
-    expect {
-      function.execute('secret/test', 'https://vault.doesnotexist:8200')
-    }.to raise_error(Puppet::Error, %r{Received 403 response code from vault at vault.doesnotexist for secret lookup.*permission denied})
+      expect {
+        function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 403 response code from vault at .* for secret lookup.*permission denied})
+    end
   end
 
-  it 'logs on, requests a secret using a token, and returns the data wrapped in the Sensitive type' do
-    connection = instance_double('Puppet::Network::HTTP::Connection', address: 'vault.doesnotexist')
-    expect(Puppet::Network::HttpPool).to receive(:http_instance).and_return(connection)
+  it 'raises a Puppet error when warning present' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupWarning)
+    vault_server.start_vault do |port|
+      expect {
+        function.execute('kv/test', "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 404 response code from vault at .* for secret lookup.*Invalid path for a versioned K/V secrets engine})
 
-    auth_response = Net::HTTPOK.new('1.1', 200, '')
-    expect(auth_response).to receive(:body).and_return(auth_success_data)
-    expect(connection).to receive(:post).with('/v1/auth/cert/login', '').and_return(auth_response)
+      expect {
+        function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
+      }.to raise_error(Puppet::Error, %r{Received 404 response code from vault at .* for secret lookup.*Invalid path for a versioned K/V secrets engine})
+    end
+  end
 
-    secret_response = Net::HTTPOK.new('1.1', 200, '')
-    expect(secret_response).to receive(:body).and_return(secret_success_data)
-    expect(connection)
-      .to receive(:get)
-      .with('/v1/secret/test', hash_including('X-Vault-Token' => '7dad29d2-40af-038f-cf9c-0aeb616f8d20'))
-      .and_return(secret_response)
+  it 'logs on, requests a kv1 secret using a token, and returns the data wrapped in the Sensitive type' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      result = function.execute('kv/test', "http://127.0.0.1:#{port}")
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('foo' => 'bar')
 
-    result = function.execute('secret/test', 'https://vault.doesnotexist:8200')
-    expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
-    expect(result.unwrap).to eq('foo' => 'bar')
+      result_opts = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
+      expect(result_opts).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result_opts.unwrap).to eq('foo' => 'bar')
+    end
+  end
+
+  it 'logs on, requests a kv2 secret using a token, and returns the data wrapped in the Sensitive type' do
+    custom_auth_segment = 'v1/custom/auth/segment'
+    vault_server = MockVault.new
+    vault_server.mount("/#{custom_auth_segment}/login", AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccessKV2)
+    vault_server.start_vault do |port|
+      result = function.execute('kv/test', "http://127.0.0.1:#{port}", custom_auth_segment, '', '', 'bar')
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('baz')
+
+      result_opts = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}", 'cert_path_segment' => custom_auth_segment, 'field' => 'bar')
+      expect(result_opts).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result_opts.unwrap).to eq('baz')
+    end
+  end
+
+  it 'is successful when providing a cert_role while authenticating' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccessWithRole)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      result = function.execute('kv/test', "http://127.0.0.1:#{port}", nil, 'test-cert-role')
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('foo' => 'bar')
+
+      opts = {
+        'vault_addr' => "http://127.0.0.1:#{port}",
+        'cert_role' => 'test-cert-role'
+      }
+      result_opts = function.execute('kv/test', opts)
+      expect(result_opts).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result_opts.unwrap).to eq('foo' => 'bar')
+    end
+  end
+
+  it 'is successful when providing a custom cert path segment without a trailing slash' do
+    custom_auth_segment = 'v1/custom/auth/segment'
+    vault_server = MockVault.new
+    vault_server.mount("/#{custom_auth_segment}/login", AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      result = function.execute('kv/test', "http://127.0.0.1:#{port}", custom_auth_segment)
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('foo' => 'bar')
+
+      result_opts = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}", 'cert_path_segment' => custom_auth_segment)
+      expect(result_opts).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result_opts.unwrap).to eq('foo' => 'bar')
+    end
   end
 
   it 'logs on, requests a secret using a token, and returns the data wrapped in the Sensitive type from VAULT_ADDR' do
-    stub_const('ENV', ENV.to_hash.merge('VAULT_ADDR' => 'https://vaultenv.doesnotexist:8200'))
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      stub_const('ENV', ENV.to_hash.merge('VAULT_ADDR' => "http://127.0.0.1:#{port}"))
+      result = function.execute('kv/test')
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('foo' => 'bar')
+    end
+  end
 
-    connection = instance_double('Puppet::Network::HTTP::Connection', address: 'vaultenv.doesnotexist:8200')
-    expect(Puppet::Network::HttpPool).to receive(:http_instance).with('vaultenv.doesnotexist', 8200, true).and_return(connection)
+  it 'logs on on Vault with a namespace, requests a secret using a token, and returns the data wrapped in the Sensitive type' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccessWithNamespace)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      stub_const('ENV', ENV.to_hash.merge('VAULT_ADDR' => "http://127.0.0.1:#{port}"))
+      stub_const('ENV', ENV.to_hash.merge('VAULT_NAMESPACE' => 'foo'))
+      result = function.execute('kv/test')
+      expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result.unwrap).to eq('foo' => 'bar')
+    end
+  end
 
-    auth_response = Net::HTTPOK.new('1.1', 200, '')
-    expect(auth_response).to receive(:body).and_return(auth_success_data)
-    expect(connection).to receive(:post).with('/v1/auth/cert/login', '').and_return(auth_response)
+  it 'caches the result when the same lookup is done more than once per catalog' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      expect(function.func).to receive(:get_secret).and_call_original.exactly(1).time
+      result1 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
+      result2 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
+      result3 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}")
 
-    secret_response = Net::HTTPOK.new('1.1', 200, '')
-    expect(secret_response).to receive(:body).and_return(secret_success_data)
-    expect(connection)
-      .to receive(:get)
-      .with('/v1/secret/test', hash_including('X-Vault-Token' => '7dad29d2-40af-038f-cf9c-0aeb616f8d20'))
-      .and_return(secret_response)
+      expect(result1).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result2).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result1.unwrap).to eq('foo' => 'bar')
+      expect(result1.unwrap).to eq(result2.unwrap)
+      expect(result1.unwrap).to eq(result3.unwrap)
+    end
+  end
 
-    result = function.execute('secret/test')
-    expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
-    expect(result.unwrap).to eq('foo' => 'bar')
+  it 'builds the cache_key correctly' do
+    vault_server = MockVault.new
+    vault_server.mount('/v1/auth/cert/login', AuthSuccess)
+    vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+    vault_server.start_vault do |port|
+      expect(function.func).to receive(:get_secret).and_call_original.exactly(3).times
+      result1 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}", 'namespace' => 'foo')
+      result2 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}", 'namespace' => 'bar')
+      result3 = function.execute('kv/test', 'vault_addr' => "http://127.0.0.1:#{port}", 'namespace' => 'baz')
+
+      expect(result1).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result2).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+      expect(result1.unwrap).to eq('foo' => 'bar')
+      expect(result1.unwrap).to eq(result2.unwrap)
+      expect(result1.unwrap).to eq(result3.unwrap)
+    end
+  end
+
+  context 'when using the agent auth method' do
+    it 'a token header is not used' do
+      vault_server = MockVault.new
+      vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+      vault_server.start_vault do |port|
+        stub_const('ENV', ENV.to_hash.merge('VAULT_ADDR' => "http://127.0.0.1:#{port}", 'VAULT_AUTH_METHOD' => 'agent'))
+        expect(function.func).not_to receive(:get_approle_auth_token)
+        expect(function.func).not_to receive(:get_cert_auth_token)
+        expect(function.func).to receive(:get_secret).with(hash_including(token: nil)).and_call_original
+        result = function.execute('kv/test')
+
+        expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+        expect(result.unwrap).to eq('foo' => 'bar')
+      end
+    end
+  end
+
+  context 'when using the agent_sing auth method' do
+    let(:agent_sink_file) { '/tmp/vault_agent_sink' }
+
+    it 'errors when token sink file does not exist' do
+      vault_server = MockVault.new
+      vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+      vault_server.start_vault do |port|
+        stub_const('ENV', ENV.to_hash.merge(
+                            'VAULT_ADDR' => "http://127.0.0.1:#{port}",
+                            'VAULT_AUTH_METHOD' => 'agent_sink',
+                            'VAULT_AGENT_SINK_FILE' => agent_sink_file,
+        ))
+        expect(function.func).not_to receive(:get_approle_auth_token)
+        expect(function.func).not_to receive(:get_cert_auth_token)
+
+        expect {
+          function.execute('kv/test')
+        }.to raise_error(Puppet::Error, %r{The agent_sink_file does not exist})
+      end
+    end
+
+    it 'a token is read from the token sink' do
+      vault_server = MockVault.new
+      vault_server.mount('/v1/kv/test', SecretLookupSuccess)
+      vault_server.start_vault do |port|
+        stub_const('ENV', ENV.to_hash.merge(
+                            'VAULT_ADDR' => "http://127.0.0.1:#{port}",
+                            'VAULT_AUTH_METHOD' => 'agent_sink',
+                            'VAULT_AGENT_SINK_FILE' => agent_sink_file,
+        ))
+        expect(function.func).not_to receive(:get_approle_auth_token)
+        expect(function.func).not_to receive(:get_cert_auth_token)
+        expect(function.func).to receive(:read_token_from_sink).with(sink: agent_sink_file).and_return('abcdefg')
+        expect(function.func).to receive(:get_secret).with(hash_including(token: 'abcdefg')).and_call_original
+        result = function.execute('kv/test')
+
+        expect(result).to be_a(Puppet::Pops::Types::PSensitiveType::Sensitive)
+        expect(result.unwrap).to eq('foo' => 'bar')
+      end
+    end
   end
 end